requirements.txt
flask==2.1.2
Jinja2==3.1.2 
Werkzeug==2.1.2
itsdangerous==2.1.2
flask-wtf
Flask-Bootstrap==3.3.7
email_validator==1.2.1
Flask-PyMongo==2.3.0
python-dotenv==0.20.0